% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_per_file.R
\name{aggregate_per_file}
\alias{aggregate_per_file}
\title{Aggregate day summary per file}
\usage{
aggregate_per_file(
  fullreport = c(),
  includenightcrit = c(),
  includedaycrit = c(),
  include24hcrit = c(),
  data_cleaning_file = c(),
  outputdir = "./",
  LUX_day_segments = c()
)
}
\arguments{
\item{fullreport}{full daysummary report from GGIR}

\item{includenightcrit}{include night criteria in hours}

\item{includedaycrit}{include day criteria in hours}

\item{include24hcrit}{include 24h window criteria in hours}

\item{data_cleaning_file}{data cleaning csv file as in GGIR}

\item{outputdir}{output directory}
}
\value{
save reports
}
\description{
Aggregate day summary per file
}
