% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGGIR.R
\name{runGGIR}
\alias{runGGIR}
\title{runGGIR}
\usage{
runGGIR(interactive = TRUE)
}
\arguments{
\item{interactive}{If TRUE (default), the directory selection is interactive}
}
\description{
Main function to run GGIR over the WHS files
}
