% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipWHSfiles.R
\name{unzipWHSfiles}
\alias{unzipWHSfiles}
\title{unzipWHSfiles}
\usage{
unzipWHSfiles(WHSfiles)
}
\arguments{
\item{WHSfiles}{directory with the files to unzip}
}
\value{
datadir definition for GGIR
}
\description{
Function to decompress the WHS files (only works with gz or 7z files)
}
